refactor: 将所有实体ID类型从Integer改为Long
统一升级项目中所有数据库实体、DTO、Repository的ID类型，解决Integer无法支持超大数据量的问题，同时更新了数据库迁移脚本适配BIGSERIAL和BIGINT类型。

+++ src/main/java/io/github/jpcndict/controller/AiPromptConfigController.java
@@ -38,7 +38,7 @@ public class AiPromptConfigController {
      * GET /api/ai-prompt-configs/{id}
      */
     @GetMapping("/{id}")
-    public AiPromptConfigVO findById(@PathVariable Integer id) {
+    public AiPromptConfigVO findById(@PathVariable Long id) {
         return service.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("提示词配置不存在，ID: " + id));
     }
@@ -67,7 +67,7 @@ public class AiPromptConfigController {
      * PUT /api/ai-prompt-configs/{id}
      */
     @PutMapping("/{id}")
-    public AiPromptConfigVO update(@PathVariable Integer id, @Valid @RequestBody AiPromptConfigRequest request) {
+    public AiPromptConfigVO update(@PathVariable Long id, @Valid @RequestBody AiPromptConfigRequest request) {
         return service.update(id, request);
     }
 
@@ -76,7 +76,7 @@ public class AiPromptConfigController {
      * DELETE /api/ai-prompt-configs/{id}
      */
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
+    public void delete(@PathVariable Long id) {
         service.delete(id);
     }
